organization of source codes are work in progress (directory compositions and function references are changed). we don't use negative infinity to calculate masked softmax anymore.
